Fix bug in Gecko RPC client and split out JsonRpcRequester to its own file
